# nvr_core/docs/swagger.yaml
basePath: /
definitions:
...
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  dto.LoginResponse:
    properties:
      permissions:
        items:
          type: string
        type: array
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/dto.LoginUser'
    type: object
  dto.LoginUser:
    properties:
      email:
        type: string
      name:
        type: string
      role_id:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
  dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshResponse:
    properties:
      permissions:
        items:
          type: string
        type: array
      token:
        type: string
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
host: localhost:9080
info:
  contact: {}
  description: API for NVR system.
  title: NVR Core API
  version: "0.1"
paths:
...
  /api/export/{cam_id}:
    get:
      consumes:
      - application/json
      description: Asynchronously exports video for a specific camera and time range.
        Returns a task ID for status polling and downloading.
      parameters:
      - description: Camera ID
        in: path
        name: cam_id
        required: true
        type: integer
      - description: Start Unix timestamp (milliseconds)
        in: query
        name: start
        required: true
        type: integer
      - description: End Unix timestamp (milliseconds)
        in: query
        name: end
        required: true
        type: integer
      - description: Stream profile (e.g., main, sub)
        in: query
        name: profile
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns task_id and status (pending)
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid camera ID or timestamps
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Missing export permissions
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Initiate a video export task
      tags:
      - Export
  /api/export/{task_id}/download:
    get:
      description: Downloads the compiled .mp4 file for a completed export task.
      parameters:
      - description: Export Task ID (UUID)
        in: path
        name: task_id
        required: true
        type: string
      responses:
        "200":
          description: The exported MP4 video file
          schema:
            type: file
        "400":
          description: Missing task ID
          schema:
            type: string
        "403":
          description: Forbidden - Missing export permissions
          schema:
            type: string
        "404":
          description: Export task not found
          schema:
            type: string
        "409":
          description: Export is not ready for download
          schema:
            type: string
        "410":
          description: Exported file has expired or been deleted
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Download exported video
      tags:
      - Export
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by username and password. Returns a short-lived
        JWT access token, a long-lived refresh token, and a list of user permissions.
      parameters:
      - description: User Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Invalid JSON payload
          schema:
            type: string
        "401":
          description: Invalid credentials or account disabled
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User login
      tags:
      - auth
  /api/logout:
    post:
      consumes:
      - application/json
      description: Invalidates a specific mobile/desktop session by revoking its refresh
        token.
      parameters:
      - description: Refresh Token to revoke
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            type: string
      summary: App Logout
      tags:
      - auth
...
  /api/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new short-lived JWT access
        token and updated permissions.
      parameters:
      - description: Refresh Token Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed
          schema:
            $ref: '#/definitions/dto.RefreshResponse'
        "400":
          description: Invalid JSON payload
          schema:
            type: string
        "401":
          description: Invalid or expired refresh token
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Refresh Access Token
      tags:
      - auth
...
  /api/web/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by username and password. Returns a short-lived
        JWT access token, a long-lived refresh token, and a list of user permissions.
      parameters:
      - description: User Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Invalid JSON payload
          schema:
            type: string
        "401":
          description: Invalid credentials or account disabled
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User login
      tags:
      - auth
  /api/web/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new short-lived JWT access
        token and updated permissions.
      parameters:
      - description: Refresh Token Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed
          schema:
            $ref: '#/definitions/dto.RefreshResponse'
        "400":
          description: Invalid JSON payload
          schema:
            type: string
        "401":
          description: Invalid or expired refresh token
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Refresh Access Token
      tags:
      - auth
swagger: "2.0"